Create Code
The code is a Python script that implements a web application for classifying images of fruits using a pre-trained machine learning model. The application is built using the Flask web framework, which provides a simple way to create web applications in Python.

The first step in the code is to import the necessary libraries. TensorFlow is used for loading the pre-trained model and making predictions, while Flask is used for creating the web application.

Next, the pre-trained model is loaded from a file using the Keras API. The model was trained to classify images of apples, bananas, and oranges, and will be used to make predictions on new images uploaded by users.

The Flask application is then defined using the Flask class. A single endpoint is defined for handling image uploads, which is /predict. When a POST request is made to this endpoint with an uploaded image, the predict() function is called.

The predict() function first gets the uploaded image from the request, and then preprocesses it using the preprocess_image() function. This function resizes the image to 224x224 pixels (which is the size expected by the pre-trained model), converts it to a numpy array, normalizes the pixels, and adds a batch dimension.

Once the image is preprocessed, the model is used to make a prediction on the image using the predict() method. The predicted class is then obtained using the get_class() function, which maps the model output to the class names.

Finally, the predicted class is returned as a JSON object using the jsonify() function, which converts the Python dictionary to a JSON format that can be sent back to the user.

The preprocess_image() and get_class() functions are utility functions that are used to preprocess the image and obtain the predicted class, respectively.

The script ends with a call to app.run(), which starts the Flask web application. If the script is run directly (as opposed to being imported as a module), the application will be started in debug mode.